+++ msmexplorer/tests/test_timeseries_plot.py
@@ -1,0 +1,14 @@
+import numpy as np
+from matplotlib.axes import SubplotBase
+
+from ..plots import plot_trace
+
+rs = np.random.RandomState(42)
+data = rs.rand(100000)
+
+
+def test_plot_trace():
+    ax, side_ax = plot_trace(data)
+
+    assert isinstance(ax, SubplotBase)
+    assert isinstance(side_ax, SubplotBase)