+++ docs/changelog.rst
@@ -14,3 +14,4 @@ New Features
 + MSM plots: ``plot_pop_resids``, ``plot_msm_network``, ``plot_timescales``
 + Projection plots: ``plot_free_energy``, ``plot_histogram``
 + TPT plots: ``plot_tpaths``
++ Misc. plots: ``plot_chord``, ``plot_trace``